requirements.txt
beautifulsoup4==4.4.1
PyYAML==3.11
requests==2.10.0
